Remove `GenericEditionPresenter`
As part of moving the rendering of content out of Whitehall, we have
introduced a Publishing API presenter for each type of document.

We therefore no longer need the `GenericEditionPresenter` that was being
used to send placeholder content items to Publishing API, instead
replacing it with a stub presenter which does the bare minimum in the
support directory.
