+++ app/models/edition.rb
@@ -426,10 +426,6 @@ class Edition < ApplicationRecord
     force_published? && state == "scheduled"
   end
 
-  def publishing_api_presenter
-    PublishingApi::GenericEditionPresenter
-  end
-
   def can_have_custom_lead_image?
     is_a?(Edition::CustomLeadImage)
   end
