+++ test/unit/app/presenters/publishing_api_presenters_test.rb
@@ -34,11 +34,6 @@ class PublishingApiPresentersTest < ActiveSupport::TestCase
                  PublishingApiPresenters.presenter_for(Speech.new).class
   end
 
-  test ".presenter_for returns a Generic Edition presenter for all models without a presenter class" do
-    assert_equal PublishingApi::GenericEditionPresenter,
-                 PublishingApiPresenters.presenter_for(GenericEdition.new).class
-  end
-
   test ".presenter_for returns a Placeholder presenter for an organisation" do
     organisation = Organisation.new
     presenter = PublishingApiPresenters.presenter_for(organisation)
